add introspection for assets and content sources (vercel/turbo#47)
add very simple explorer to introspect the graph via `/__turbopack__/`

Allows to "walk" through the content sources and asset graph. Some
assets allows to access inner assets, e. g. from the ecmascript chunk
you can access the entry module asset, so you can go a level deeper.

It works by a `Introspectable` trait which can be implemented by
anything we want to allow the user to inspect. In future the in browser
UI should offer a nice UI to explore your application.

In future we probably also want to add some `metrics` to the
`Introspectable` trait, e. g. to report size of an asset.

This UI is very basic currently, but that works for debugging:


![image](https://user-images.githubusercontent.com/1365881/196007862-2fb2a0e9-19ba-45c5-8a82-926418e0d479.png)