Always mark styled-jsx as external to avoid duplicating with bundled copies (vercel/turbo#3386)
Fixes WEB-412.

Previously, references to styled-jsx inserted by the swc transform would be bundled, which duplicated the copy required by next, which is external. This marks the dependency is to be resolved as external always.

Test Plan: Verified in a small app that `<style jsx>` nodes result in a server-rendered response with the appropriate `<style>` tag in the head.